ENGR00229188 imx6 sabresd: keep the w/h size ratio for thumbnail
Keep the width/height ratio for activity screenshot thumbnail.
The original height of the thumbnail is 135dp, which
cause am use the height low ratio to get screenshot with
size of 199x144. This screenshot is too small to fit
the status bar's recent activity panel view.
Increase the height size for thumbnail by size of
225x170, which fit the image view and avoid blank
area on the right of the app thumbnail.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>